% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fetch.data.R
\name{ERAI.fetch.data.for.month}
\alias{ERAI.fetch.data.for.month}
\title{Get the data for one variable for one month}
\usage{
ERAI.fetch.data.for.month(var, year, month)
}
\arguments{
\item{var}{- 20CR variable name}
}
\description{
Use 20CR variable names, store the data on $SCRATCH
}
