% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fetch.data.R
\name{ERAI.fetch.forecast.data.for.month}
\alias{ERAI.fetch.forecast.data.for.month}
\title{Get data for one forecast variable}
\usage{
ERAI.fetch.forecast.data.for.month(var, year, month)
}
\arguments{
\item{var}{- 20CR variable name}
}
\description{
I'm packaging data by the month, a compromise
  between speed, space, & simplicity.
}
\details{
Won't replace data already on disc - just
  returns immediately if the result file is already there.
}
