% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GSDF.ERAI.R
\name{ERAI.get.data.dir}
\alias{ERAI.get.data.dir}
\title{ERAI get data directory}
\usage{
ERAI.get.data.dir()
}
\value{
Base directory name (or NULL is no local files)
}
\description{
Find local data directory - different for different systems
}
\details{
The various different systems on which I run this code all have
 different places to keep large data files. This function returns
 the right base directory for the system (if there is one).
}

