% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSDF.ERAI.R
\name{ERAI.get.slice.at.hour}
\alias{ERAI.get.slice.at.hour}
\title{Get slice at hour.}
\usage{
ERAI.get.slice.at.hour(variable, year, month, day, hour, height = NULL)
}
\arguments{
\item{variable}{'prmsl', 'prate', 'air.2m', 'uwnd.10m' or 'vwnd.10m' - or any supported variable.}

\item{height}{Height in hPa - leave NULL for monolevel}
}
\value{
A GSDF field with lat and long as extended dimensions
}
\description{
Get a 2D horizontal slice of a selected variable (as a GSDF field) for a given hour.
}
\details{
Interpolates to the selected hour when the data available are less than hourly.
Interpolates to the selected height when the selected height is not that of a ERA5 level.
}

