% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GSDF.ERAI.R
\name{ERAI.translate.for.variable.names}
\alias{ERAI.translate.for.variable.names}
\title{Translate to ERA Interim variable names in the .nc files}
\usage{
ERAI.translate.for.variable.names(var)
}
\arguments{
\item{var}{- 20CR variable name}
}
\value{
name used for ERAI variable names
}
\description{
I'm standardising on 20CR variable names - map to ERAI choices
}

