% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSDF.ERAI.R
\name{ERAI.hourly.get.file.name}
\alias{ERAI.hourly.get.file.name}
\title{ERAI get file name (hourly)}
\usage{
ERAI.hourly.get.file.name(variable, year, month, day, hour)
}
\arguments{
\item{variable}{'prmsl', 'prate', 'air.2m', 'uwnd.10m' or 'vwnd.10m' - or any supported variable}
}
\value{
File containing the requested data
}
\description{
Get the file name for selected variable and date (hourly data)
}
\details{
Called internally by \code{ERAI.get.slice.at.hour} but also useful
 called directly - you can then access the data with another tool.
}

